istqb_trainer/requirements.txt
pdfplumber>=0.11.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.40.0
anthropic>=0.34.0
